requirements.txt
google-cloud-aiplatform
qdrant-client
python-dotenv
langchain
langchain-community
langchain-qdrant
langchain-huggingface
flask
google-generativeai
jsonify
twelvelabs
google-cloud-texttospeech
grpcio
langchain_google_genai
flask_cors